TFale-main/flask-app/requirements.txt
flask
psycopg2-binary